backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
python-dotenv>=1.1.0,<1.2.0
pydantic>=2.11.0,<3.0.0
platformdirs>=4.3.0,<5.0.0
pyyaml>=6.0.2,<7.0.0
httpx>=0.28.0,<0.29.0
python-multipart>=0.0.7,<0.1.0 
openai>=1.56.0,<2.0.0   